requirements.txt
python-docx>=1.1
openpyxl>=3.1
pywin32>=306
Pillow>=10.0
requests>=2.31
